Add a quit application event handler for back button, move QR code to manufacturer activity
