+++ app/src/main/java/com/example/barcode_a/Home_Manufacturer.kt
@@ -8,19 +8,18 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.LinearLayout
+import android.widget.Toast
+import com.google.firebase.auth.FirebaseAuth
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-/**
- * A simple [Fragment] subclass.
- * Use the [Home_Manufacturer.newInstance] factory method to
- * create an instance of this fragment.
- */
+
 class Home_Manufacturer : Fragment() {
-    // TODO: Rename and change types of parameters
+
+    private lateinit var firebaseAuth: FirebaseAuth
+
     private var param1: String? = null
     private var param2: String? = null
 
@@ -33,6 +32,19 @@ class Home_Manufacturer : Fragment() {
     }
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        firebaseAuth = FirebaseAuth.getInstance()
+
+        val signOut = view.findViewById<Button>(R.id.btnSignout)
+        val activity = requireActivity()
+
+        //Sign Out Function
+        signOut.setOnClickListener(){
+            firebaseAuth.signOut()
+            Toast.makeText(activity , "Account Signed Out!" , Toast.LENGTH_SHORT).show()
+            val intent = Intent(activity, LoginTab::class.java)
+            startActivity(intent)
+        }
 
         val layoutProductInfo = view.findViewById<LinearLayout>(R.id.layoutProductInfo)
         layoutProductInfo.setOnClickListener {
