+++ app/src/main/java/com/example/barcode_a/MainActivityManufacturer.kt
@@ -2,15 +2,21 @@ package com.example.barcode_a
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.os.Handler
+import android.widget.Toast
 import com.example.barcode_a.databinding.ActivityMainBinding
 import androidx.fragment.app.Fragment
 import com.example.barcode_a.databinding.ActivityMainManufacturerBinding
+import com.google.firebase.auth.FirebaseAuth
 
 
 class MainActivityManufacturer : AppCompatActivity() {
 
     private lateinit var binding: ActivityMainManufacturerBinding
 
+
+    private val delay : Long = 3000 // 3 seconds delay
+    var quit = false
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -20,6 +26,7 @@ class MainActivityManufacturer : AppCompatActivity() {
         setContentView(binding.root)
 
         replaceFragment(Home_Manufacturer())
+
         binding.bottomNavigationView.setOnItemSelectedListener {
 
             when(it.itemId){
@@ -39,6 +46,23 @@ class MainActivityManufacturer : AppCompatActivity() {
         val fragmentTransaction = fragmentManager.beginTransaction()
         fragmentTransaction.replace(R.id.frame_layout2,fragment)
         fragmentTransaction.commit()
+    }
+
+    //Quit Application
+    override fun onBackPressed() {
+
+        if (quit == false){
+            Toast.makeText(this, "Press Again To Quit", Toast.LENGTH_SHORT).show()
+            quit = true
+
+            val handler = Handler()
+            handler.postDelayed({
+                quit = false
+            }, delay)
+        }
+        else{
+            finishAffinity()
+        }
     }
 
 }
