+++ app/src/main/java/com/example/barcode_a/Search.kt
@@ -1,37 +1,25 @@
 package com.example.barcode_a
 
-import android.graphics.Bitmap
-import android.graphics.Color
-import android.graphics.ImageFormat.RGB_565
-import android.media.Image
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.EditText
-import android.widget.ImageView
-import android.widget.Toast
-import com.budiyev.android.codescanner.AutoFocusMode
-import com.budiyev.android.codescanner.CodeScanner
-import com.budiyev.android.codescanner.CodeScannerView
-import com.budiyev.android.codescanner.ScanMode
-import com.google.zxing.BarcodeFormat
-import com.google.zxing.WriterException
-import com.google.zxing.qrcode.QRCodeWriter
 
-
+// TODO: Rename parameter arguments, choose names that match
+// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
+/**
+ * A simple [Fragment] subclass.
+ * Use the [ProductInformation.newInstance] factory method to
+ * create an instance of this fragment.
+ */
 class Search : Fragment() {
-
+    // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-    private lateinit var ivGenerateQR : ImageView
-    private lateinit var etGenerateEnterData : EditText
-    private lateinit var btnGenerateGenerateQR : Button
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -41,52 +29,27 @@ class Search : Fragment() {
         }
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        val activity = requireActivity()
-
-        ivGenerateQR = view.findViewById(R.id.ivGenerateQR)
-        etGenerateEnterData = view.findViewById(R.id.etGenerateEnterData)
-        btnGenerateGenerateQR = view.findViewById(R.id.btnGenerateGenerateQR)
-
-        btnGenerateGenerateQR.setOnClickListener(){
-            val data = etGenerateEnterData.text.toString().trim()
-
-            if (data.isEmpty()){
-                Toast.makeText(activity, "Enter some data", Toast.LENGTH_SHORT).show()
-            }else{
-                val writer = QRCodeWriter()
-                try {
-                    val bitMatrix = writer.encode(data, BarcodeFormat.QR_CODE, 512,512)
-                    val width = bitMatrix.width
-                    val height = bitMatrix.height
-                    val bmp = Bitmap.createBitmap(width, height,Bitmap.Config.RGB_565)
-                    for (x in 0 until width){
-                        for (y in 0 until height){
-                            bmp.setPixel(x, y, if(bitMatrix[x, y]) Color.BLACK else Color.WHITE)
-                        }
-                    }
-                    ivGenerateQR.setImageBitmap(bmp)
-
-                }catch (e: WriterException){
-                    e.printStackTrace()
-                }
-            }
-        }
-    }
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+        // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_search, container, false)
     }
 
     companion object {
-
+        /**
+         * Use this factory method to create a new instance of
+         * this fragment using the provided parameters.
+         *
+         * @param param1 Parameter 1.
+         * @param param2 Parameter 2.
+         * @return A new instance of fragment ProductInformation.
+         */
+        // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            Search().apply {
+            ProductInformation().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
