+++ app/src/main/res/layout/fragment_home__manufacturer.xml
@@ -145,6 +145,18 @@
 
             </LinearLayout>
 
+            <Button
+                android:id="@+id/btnSignout"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Sign Out"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintHorizontal_bias="0.803"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintVertical_bias="0.813" />
+
         </androidx.constraintlayout.widget.ConstraintLayout>
     </ScrollView>
